+++ src/main/java/cart/dao/dummyData/ProductInitializer.java
@@ -1,7 +1,7 @@
 package cart.dao.dummyData;
 
 import cart.dao.ProductDao;
-import cart.entity.ProductEntity;
+import cart.domain.entity.ProductEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
