+++ src/main/java/cart/domain/entity/ProductEntity.java
@@ -1,4 +1,4 @@
-package cart.entity;
+package cart.domain.entity;
 
 import cart.domain.Price;
