+++ src/main/java/cart/service/MemberManagementService.java
@@ -2,7 +2,7 @@ package cart.service;
 
 import cart.dao.MemberDao;
 import cart.dto.MemberDto;
-import cart.entity.MemberEntity;
+import cart.domain.entity.MemberEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
