+++ src/test/java/cart/dao/JdbcProductDaoTest.java
@@ -1,7 +1,7 @@
 package cart.dao;
 
 import cart.dao.dummyData.ProductInitializer;
-import cart.entity.ProductEntity;
+import cart.domain.entity.ProductEntity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
