+++ src/test/java/cart/service/MemberManagementServiceTest.java
@@ -2,7 +2,7 @@ package cart.service;
 
 import cart.dao.JdbcMemberDao;
 import cart.dto.MemberDto;
-import cart.entity.MemberEntity;
+import cart.domain.entity.MemberEntity;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
